front/src/app/features/auth/components/login/login.component.ts
import { Component, NgZone } from '@angular/core';
import { FormBuilder, Validators } from '@angular/forms';
import { Router } from '@angular/router';
import { SessionInformation } from 'src/app/interfaces/sessionInformation.interface';
import { SessionService } from 'src/app/services/session.service';
import { LoginRequest } from '../../interfaces/loginRequest.interface';
import { AuthService } from '../../services/auth.service';

@Component({
  selector: 'app-login',
  templateUrl: './login.component.html',
  styleUrls: ['./login.component.scss'],
})
export class LoginComponent {
  public hide = true;
  public onError = false;

  public form = this.fb.group({
    email: ['', [Validators.required, Validators.email]],
    password: ['', [Validators.required, Validators.min(3)]],
  });

  constructor(
    private authService: AuthService,
    private fb: FormBuilder,
    private router: Router,
    private sessionService: SessionService,
    private ngZone: NgZone // Ajoutez NgZone comme dépendance du constructeur
  ) {}

  public submit(): void {
    const loginRequest = this.form.value as LoginRequest;
    this.authService.login(loginRequest).subscribe({
      // utiliser la méthode ngZone.run() pour exécuter la navigation à l'intérieur de la zone Angular
      next: (response: SessionInformation) =>
        this.ngZone.run(() => {
          this.sessionService.logIn(response);
          this.router.navigate(['/sessions']);
        }),

      error: (error) => (this.onError = true),
    });
  }
}
